separate quick punch
